MacScribe:  Add quad<->string conversions
The term 'quad' is just our generic name for FourCharCode or OSType.  These
functions convert an arbitrary 4-byte code into a suitable filename (taking
into account illegal characters like NUL and '/' and others that are unsafe
for certain purposes) and parse the resulting name back into a quad.

For now, we take a conservative tack on character safety and allow only '_'
and alphanumerics (making an exception for '#' in the last position).  As a
special case, up to three spaces consecutive trailing spaces are considered
safe, in the sense that stripping them has no effect, since the parser will
assume that the missing characters are spaces and append enough to get four
bytes.