v68k:  Prefetch instructions
Prefetch an instruction word in reset(), so that a bogus PC causes a double
bus fault.  Also prefetch an instruction at the end of a successful step(),
and at the beginning, merely advance the PC.

This will also simplify the implementation of BKPT, which asks the hardware
for a replacement instruction word.