v68k-test:  Write "Hello world" for real
Actually write "Hello world" to the console instead of settling for "HI" as
hexidecimal art in the register dump.

We use a very simple API:  Breakpoint vector 0 is used to write a character
(passed in D0) to standard out.  We could treat D0 as a Unicode code point,
but for now we'll limit ourselves to ASCII.  If the character value exceeds
the range of ASCII, then an illegal instruction exception will occur at the
next step() call, due to the unacknowledged breakpoint.