v68k:  Add load/store instruction flags
Define instruction flags for specifying the size of stored data, whether it
also needs to be loaded from the destination as input to the operation, and
whether the destination is a register vs. memory.

Overload operator|() so we can write loads_and|stores_word_data|in_register
(for example) and still evaluate as instruction_flags_t.