v68k:  Add load/store mechanism
Factor out code to load and store destination operand data so it needn't be
duplicated among microcode functions.

The new code in step() (with the help of load() and store() functions)

* yields an address error on badly aligned word or longword access
* yields a bus error if the load or store fails
* provides the loaded data to the microcode in params[1]
* allows the microcode to simply assign the result back to params[1]
* masks the result appropriately when the destination is a data register
* sign-extends the result when the destination is an address register
* sets the CCR bits appropriately (see below)

The CCR is set as follows when the destination is not an address register:

* X is unchanged
* N is set if the result is negative, and cleared otherwise
* Z is set if the result is zero, and cleared otherwise
* V is always cleared
* C is always cleared

This is the most common behavior.  When we have instructions with different
requirements (like ADD and SUB), we'll need a way to suppress this, such as
a new flag.