v68k:  Implement the Vector Base Register
Set it to zero in reset().  Add it to the vector offset before fetching the
vector.

The 68000 doesn't have a VBR (and doesn't implement MOVEC), so the internal
value will always be zero.  The 68020 and later don't fault on odd longword
accesses.  But to support the 68010 we must check for odd VBR values.