v68k:  Add fetches_MOVEM
Fetch the size code.  We only examine the low bit, but it's simpler to have
the filter in the microcode than write a new fetcher.

Fetch the sized effective address.  This one has to be special to MOVEM due
to the custom size code.

Fetch the update register id / flag.

Fetch the register mask.