v68k-exec:  Provide just enough MacRelix context
Allocate the user and system parameter blocks (statically), give the latter
a pointer to the former, and pass the latter (with argc, argv, and envp set
to zero) to the loaded code.  MacRelix tools will try to write to fields in
the user parameter block, whose address is loaded from the system parameter
block.