which:  Avoid strings, allocate memory just once
Traverse the directories in PATH and the programs in argv and determine the
longest of each.  Allocate a buffer large enough to hold any combination of
directory and program.  For each program, write the program name at the end
of the buffer and each directory immediately preceeding it, so we only need
to copy the program name once.  Allow the newline to overwrite the C string
terminator since we don't need it for write().