v68k-exec:  Expand and reorganize memory
Double the memory size to 16K.  The user code area still occupies the upper
half (which is now 8K).  The 4K previously housing the user code is now for
parameter storage, containing the user and system parameter blocks, as well
as argc and argv.  Since this area is not shared with a stack, an arbitrary
cut-off limit isn't needed, and the max size applies to all parameter data,
not just argv.

Reverse the order of the user and supervisor stacks.  The latter now shares
a 1K block with the OS, while the former gets a full 2K.