v68k:  Avoid missing MULU instruction; use AND
Instead of multiplying a value by a factor of 0 or 1, AND it with 0 or -1,
since multiplication instructions aren't implemented yet.

Also, it's probably more efficient.

This is needed so v68k-hello can run emulated in v68k-exec.

Multiplications of power-of-2 constants are already replaced by bit shifts
and don't need to be changed.