Revert "v68k:  Fix CLR to read first on 68000"
The patch is broken, and irreparably so.  The idea of delegating CLR either
to AND or MOVE would work if decoding had access to the selected processor,
but we've deliberately chosen to avoid passing machine state to the decoder
at runtime or baking in the choice at compile time.  The offense happens in
load(), so we have to deal with it there.

This reverts commit 0b93975c6ed63cae4a5d441c475a9aacba3f4211.