v68k:  Rename and_sets_CCR to basic_CCR_update
Previously, a flag set might be spelled loads_and|stores_data|and_sets_CCR,
which worked because stores_data was a prerequisite for the other two.  But
the instructions TST and CMP update the CCR without storing anything, so we
dispense with the linguistically cute 'and' prefix.  In addition, there are
different kinds of CCR updating, so we specify the most common one 'basic'.