[DEL] v68k-bitshift:  Purge superfluous v68k tests
These tests explicitly using v68k emulator instances have been obsoleted by
shift.68k, which must be compiled for 68K and run in v68k-exec.  While this
means that you can't rely on the native compiler to build the tests, on the
plus side, the tests are far easier to write, *and* they can be run without
emulation for the purpose of validating the tests themselves.