PosixLib:  Add errno
Define errno here, so it will be consistently ignored in both MSL C.PPC.Lib
and OTXTILib.  Otherwise, it will be defined by whichever of those is first
in the link order, and since there's no requirement either way, it's better
not to depend on it.  (Case in point:  The order changed in the recent conf
file cleanup, yielding a new link warning; this patch and the next fix it.)