[FIX] Genie:  Properly recreate sys console window
Remove the main view if it exists before trying to instantiate it.

Closing the System Console window removes the window, but leaves behind the
port in which the views are instantiated.  Consequently, the link operation
to instantiate a scrollframe fails and an exception traverses MakeWindow().

It would be convenient to leave the view hierarchy in place, but at present
there's no way to set the fonts correctly in this case -- the TextEdit view
is installed when the window is created, at which time it inherits its font
properties from the window, but we can't alter those properties until after
the window is created.  One solution is to allow QuickDraw port settings to
be established in advance, the way arguments to NewWIndow() are.  A cleaner
approach is to use the window's font as a default, but allow TextEdit views
to override it with its own text-* property files.