Genie:  Mark processes active after exec
This was originally done during the port to Mach-O, but that issue seems
to have worked itself out.  However, this change is useful in System 7.1
with the Thread Manager extension:  It averts a stall during application
launch (whereby MacRelix launches, but doesn't create a window until the
first event is received).

Unfortunately, while the very same issue occurs with our custom threads,
this change has no observable effect.