Genie:  Avoid frame pointer missing-return warning
Calling recall::get_frame_pointer() yields a warning that control may reach
the end of the non-void function.  The function must return a value (and be
declared as such), but the only code is in assembler.  While we can declare
the entire function 'asm' when targeting Metrowerks (such that the compiler
refrains from this warning due to lacking the logic to analyze the assembly
code), gcc's __asm__() construct to insert assembler instructions is like a
self-contained asm block, in that that the function body otherwise contains
C (or C++) code and is expected to conform appropriately.  To fully address
this will require fixing the inline assembler usage, but for now we silence
the warning by avoiding use of get_frame_pointer() outside of Metrowerks in
the first place.  (The stack limit isn't currently being used anyway.)