Rename x68k to xv68k
The x68k name is overloaded on the Web to the point that no reference to
v68k occurs within the first two pages of results from either DuckDuckGo
or Google.  On the other hand, Google reports fewer than 250 results for
xv68k -- taking the top slot should be no problem.  A further benefit is
that xv68k contains the literal string 'v68k' while still beginning with
'x' (for "execute").