Genie:  Don't call async FSpOpenDF()
Somehow this fails on 68K.  Just avoid it.  Call N::FSOpenDF() instead.

On an initial run of MacRelix with the perl library already present, the
call to symlink() would successfully create the alias but throw OSStatus
1 instead of writing the link target to the data fork (and keep the file
open).  This is now fixed.

It seems like trying to do asynchronous file I/O in Mac OS is a waste of
time.