Genie:  Fix leak of view vfs nodes
We were mistakenly passing object pointers to simple_map::erase(), which
unlike std::map takes only keys, not iterators, and the type-checking is
broken such that any pointer is accepted, so we were trying to erase the
wrong pointer, which had no effect.

In RemoveViewParameters(), this would only leak an empty submap, whereas
in RemoveAllViewParameters(), the entire set of subviews (of which there
is invariably one or zero, except for stacks) would remain in the submap
under the parent's key.  Since views maintain strong references to their
parent nodes, the port nodes remain in /gui/port as well.

gViewParametersMap eventually gets destroyed after main() returns, which
has been the source of much trouble when destructors refer to structures
which no longer exist.