Genie:  Add layer nodes between stack and views
Insert non-view layer nodes between a stack view and its subviews.  This
changes the subview path from (e.g.) view/foo to view/foo/v (where "foo"
is the unique name identifying the subview in the stack).  To allow this
path to be passed directly to link(), view/foo (for any foo) will always
resolve to an existing directory (although it only persists and shows up
in a directory listing if the subview is instantiated).  This will allow
"v" to be split into "v" and "view", of which the first is the directory
and the second is the file created by link(), for compatibility with the
POSIX semantics enforced by FUSE.