Genie:  Allow views nested within non-views (once)
In unview_mkdir(), if a view's parent does not have an associated window
(because it's not itself a view), try the grandparent.

Layered stacks break the assumption that each subview's parent is itself
a view.  An unlayered stack's subviews were its direct children, so this
held true, but an unlayered stack's children (and its subviews' parents)
are layers, which are not views.