Genie:  Remove magic /app filesystem
The /app directory stores symbolic links that are magically removed when
their respective targets cease to exist and they would otherwise dangle.

This was used for FORGE applets for which it was desirable to run only a
single instance at a time, such as the About box.  The About application
would create a link in /app to its port (in /gui/port) which a later-run
instance would detect (and thereby avoid creating two About boxes on the
screen simultaneously).  Since about and psmon now use the new lock file
instead, /app is now obsolete.