mwcc:  Omit -w noimplicit on 68K
CodeWarrior Pro 3 doesn't recognize noimplicit, so we have to omit it in
order to compile anything.  CodeWarrior Pro 6 does recognize it and will
warn about implicit arithmetic conversions if '-w all' isn't followed by
it.  Since the only benefit of Pro 3 over Pro 6 is that it's 68K-hosted,
we can assume that Pro 3 users have 68K systems and those with PPC would
prefer Pro 6.  Therefore, omit noimplicit on 68K to avoid compile errors
with Pro 3 and include it on PPC to avoid the warnings with Pro 6.  This
compromise also allows the warnings to be generated by building with 68K
MacRelix and CW Pro 6.