Genie:  Make port's "w" a real directory in Mach-O
In MacRelix, the "w" file in a port's directory is a symlink to a window
directory, "/sys/app/window/list/$address".  If we're exporting the user
interface over a network, it's a bad idea to include /sys, which enables
access to disk filesystems.  Instead, make "w" a real directory if we're
building in a configuration that may be used as a display server instead
of running user programs locally.