__forktool:  Call main() on new, mmap-backed stack
Call _relix_restack() to create a new stack on which to run main().  The
new stack is allocated using malloc(), which gets its backing store from
mmap()[1].  Since malloc's memory uses the MAP_PRIVATE flag, fork() will
keep distinct copies of the stack in parent and child.  Though the stack
originally provided by the Thread Manager remains shared, this is a non-
issue, since exit() is called on the new stack and control never returns
to the initial stack.

[1] Actually, malloc() is backed by mmap() only if dlmalloc is used, and
otherwise by NewPtr() via MSL.  But the idea of this patch is to support
fork(), which implies dlmalloc already.  Programs that don't call fork()
may continue using MSL's malloc(), and the stack will be cleaned up when
static destruction occurs.