Genie:  Use app location as root dir if no "/"
If the "/" directory is not found in the app's location, then instead of
seeking it in the volume's root, just use the app's location as the root
directory.

In Mac OS X, the second GetCatInfo() call throws nsvErr if vRefNum == 0,
whereas Mac OS 9 throws fnfErr if vRefNum is assigned the actual number.
To skirt this issue, call GetProcessBundleLocation() for Carbon to cover
Mac OS X (which also works for Mac OS 9), and otherwise code for classic
Mac OS.