UseEdit:  Silence clang foreign enum value warning
When we use "case typeDocument" in a switch of type Mac::DescType, clang
warns, "case value not in enumerated type".  While I'm not ready to rule
out permitting a translation unit to add custom definitions to DescType,
that would require significant forethought.  Instead, just switch on the
integer type ::DescType.