tests:  Remove test-longjmp-past-vfork tests
The test for a child of vfork() that exits or execs after returning from
the scope in which vfork() was called is not compatible with system call
stacks.  The in-kernel check needs to be rewritten to allow for the call
to vfork() and the exec or exit occuring on different stacks.  (In fact,
this was already an issue, but one that went unnoticed.)  Until then, we
proceed without such a check, and tests relying on it are invalid.