teletype:  Set $PORT as well as $WINDOW
This is part of a terminology transition from 'window' to 'port'.  While
X11 uses windows which are either 'mapped' or not, FORGE holds ports and
windows as distinct concepts.  While there'll usually be one window to a
port, it's conceivable that in the future it will be possible to display
the same port's contents in different windows simultaneously -- as would
be useful for a slide presentation, where the slides display full-screen
on a projector, *and* in a window on the control terminal -- in addition
to windows offering an outline view, thumbnails, and controls.