PosixLib:  Recalibrate clock when lag hits 2s
To get microsecond resolution, gettimeofday() relies on a mashup of both
the system clock and the Microseconds() timer.  However, the micro timer
measures hardware uptime, and doesn't advance during system sleep, while
the system clock does.  Therefore, it's necessary to check for "jet lag"
and recalibrate the offset between the micro timer and global Unix time.

We have to apply a threshhold instead of updating every time, or we'd be
right back to one-second resolution.