v68k-alloc:  allocate() for existing native memory
Factor out new function allocate_n_pages_for_existing_alloc_unchecked(),
which takes a page count and a pointer to a native memory block matching
in size.

Add an inline two-arg allocate() that takes a native pointer plus a byte
count and calls allocate_n_pages_for_existing_alloc_unchecked() -- after
some computation which can be optimized away when the value of 'size' is
known at compile-time.  The size of the memory must be a multiple of the
page size, because the allocator just uses whole pages and doesn't store
the actual allocation size anywhere, so an entire page must be backed by
real memory.