v68k-mac:  Fix trap dispatcher for A0 arguments
We were always clobbering A0, which is used to pass arguments to some OS
traps.  To fix:  Save A2, and don't touch it for Toolbox traps (since A0
works just fine), and restore it after using it to hold the trap address
for OS traps (and also restoring the original A0 before the jump).