Genie:  Make Process_ThreadEntry() a free function
Move Process::ThreadEntry() out of the class and the header.  It remains
extern because it's used as a template parameter.

This is important because in Apple's gcc, 'pascal' is a macro instead of
a keyword, and a Mac OS header must be included to define it, which we'd
rather do outside of our own headers.