Genie:  Rewrite pump() to avoid Peek()
One consequence is that a write failure may result in data getting lost,
but I don't think we were depending on unwritten data being subsequently
readable, and I'm not convinced pump() was working properly in all cases
anyway.  In any event, non-destructive reads need to be rearchitected to
work directly with pipe and paired socket buffers, and to reuse whatever
means we come up with for MSG_PEEK with other sockets.