Genie:   Don't repeat fd get logic in select()
Call relix::get_fd_handle() instead of duplicating its logic.  This will
run more instructions, since the fd_map fetch is done each time, but the
whole point of select() is to wait for something, so this shouldn't have
a negative effect on performance.  Also, it produces smaller object code
and reduces coupling.