vfs:  Add map_uninitialized()
Use temporary memory for requests larger than 64K (or if malloc() fails)
on Mac OS.  Otherwise, just use malloc().

The map_temporary() function is defined in MacVFS.  Since MacVFS isn't a
prerequisite of vfs (and it can't be, because it's a dependant), linking
to it is non-portable -- i.e. it won't work with a Unix toolchain -- but
it's only required for classic Mac OS anyway, and Metrowerks handles it.