Genie:  Reset process attrs for main thread only
Don't reset the shared process attributes (process resources and process
image) when terminating a non-leader thread.  Doing so when signalling a
process group result in a wait() on a process reporting an exit of zero,
rather than death via signal, if the process has created a pthread whose
task is terminated before the process' task.