[FIX] Genie:  Set a usable file on console handles
With the existing shim object, `cat /dev/fd/0` and `echo eh > /dev/fd/1`
will fail.  Replace it with a real file (obtained from a path resolution
which depends on the dynamic element, which requires the tty filehandle,
which is why we don't just pass it to filehandle() in the first place).