Genie:  Simplify Mac HFS fork-opening code
Replace New_DataForkHandle() and New_RsrcForkHandle(), which differ only
in which FileGetter they pass to MacFileHandle(), with a single function
that takes a FileGetter argument.  Pass that to OpenMacFileHandle() from
hfs_open() and rsrcfile_open().  This reduces duplicate code and removes
the cyclical dependency between the FSTree_RsrcFile and MacFile units.