[NEW] display-fb:  Paint raster images via linuxfb
This works splendidly on a Raspberry Pi.  On a 2005 consumer PC, it also
works, with the caveat that the internal graphics only supports 800x600.
On a 2013 rig custom-built with cheap parts, it supports a limited array
of resolutions, only drawing correctly at 32-bit depth, and even then at
an overly wide aspect ratio.