mac:  Handle left/right masking in the same byte
If the left edge is right of a byte boundary, and the right edge is left
of the next byte boundary, it violates our assumption that the number of
(horizontal) edges on non-byte boundaries (in this case, two) equals the
number of masked pixels (in this case, one).  In other words, we assumed
that there would be zero or more skipped pixels on the left, then either
a byte boundary or a masked pixel, then zero or more copied pixels, then
either a byte boundary or a masked pixel, and finally the skipped pixels
on the right.  In this case, the left and right masks are applied to the
same pixel, and therefore the inner pixel range (starting after the left
masked pixel and ending before the right one) actually runs backward (by
one pixel), which results in attempting to pass -1 to memset(), with the
outcome you'd expect.