Carbonate:  Compile Carbon accessors on PPC also
Instead of linking against CarbonAccessors.o on PPC, reuse the functions
we already wrote for use with 68K (if ACCESSOR_CALLS_ARE_FUNCTIONS isn't
set).  (Suppress the new link warnings since we haven't actually trashed
CarbonAccessors.o yet.)

This is necessary to unset ACCESSOR_CALLS_ARE_FUNCTIONS:  If we had just
unset it, we'd get a linker warning complaining that GetNextWindow() was
redefined in CarbonAccessors.o (though I haven't the slightest idea how,
since the other definition is inline).  If we jettison CarbonAccessors.o
on the other hand, we'll need definitions of any functions which are now
68K-only.  As a bonus, this eliminates the nuisance of CodeWarrior IDE's
irritating "This source file contains null characters.  Are you sure you
want to open it?" alert *and* lets you actually step into source code in
the debugger.