mac-types:  Add VRefNum_DirID structure type
I wasn't sure what to call this.  Nitrogen has the name FSDirSpec, which
is reasonable enough but doesn't convey that it's a third-party addition
rather an original member of the namespace.  Rather than inject the most
Toolboxish name I can think of as an intended equal among its neighbors,
I've opted for austerity.  The name VRefNum_DirID tells you exactly what
the struct contains and nothing more.  (And the underscore is a giveaway
that it's not part of the Toolbox.)