tools:  Axe EXIT_{SUCCESS,FAILURE}
The definitions are fixed, so there's no abstraction benefit.  Often, it
necessitates including <stdlib.h> when it's otherwise unneeded.  And the
inconsistency in how forgiving various standard libraries are of missing
includes creates an ongoing risk that code that builds in one place will
fail in another, such as OS X 10.9.