Nitrogen iteration:  Store AEDescList in iterator
Store a copy of the AEDescList in AEDescList_ItemData_ValueIterator, not
just a reference.  This silences a warning about the default constructor
(which was clearly wrong).  As for whether it makes the intended use any
safer, it removes the possibility of referencing the AEDescList after it
ceases to exist, but if the calling code calls AEDisposeDesc() (and sets
the dataHandle to NULL) on the original, then the iterator will possibly
crash calling AEGetNthPtr() on its stale copy, instead of getting a mere
error.  In any event, correct usage doesn't change, and this patch fixes
the default case.