Genie:  On Quit, send SIGKILL to stopped threads
Sending SIGTERM doesn't affect a stopped process (until a SIGCONT signal
is received, which doesn't tend to happen automatically).

If some program relies on catching SIGTERM so it can send SIGCONT to its
stopped coprocess, which also catches SIGTERM (so it can clean up before
exiting), then this will need to be revisited.