Genie:  Fix rename "foo" -> "Foo"
The duplicate name detection was broken.  The case-altered name spelling
appeared as an existing destination item, which was then deleted to make
way for the rename, which then failed since the source no longer existed
(having just been deleted).

Instead of trying to predict in advance whether the rename will fail due
to an existing item, just try it (and check for dupFNErr), in which case
delete the item and retry the rename.  This future-proofs us against the
prospect of a case-sensitive filesystem (e.g. HFSX) as well as any other
scenario in which the traditional File Manager's idea of string equality
doesn't hold.