[NEW] Beep:  Simple Mac application that beeps
Initialize the menus so the menubar will flash if sound is muted.  Delay
afterward to let the sound finish so it doesn't get cut off upon exiting
the application.  (Except skip InitWindows() and InitMenus() when run in
xv68k, which doesn't implement those yet and doesn't need them to beep.)