rom-copier:  Just give up if the ROM file exists
The original idea was that the user could move or delete the file before
continuing, but the most likely scenario is a prior successful ROM dump,
in which case no further action is useful.  Another possible scenario is
a truncated ROM image from an interrupted dump, in which case a checksum
(MD5 at least) won't match.  A future patch should check an existing ROM
file in addition to newly created ones.