[NEW] fifo-signal:  Unblock fifo-waiters
Unlike fifo-wait, which could be called from a shell script, this is not
actually usable in a server as is.  A server using this metaprotocol has
to keep the FIFO open for the benefit of clients which open their end of
the FIFO after the server starts.  (While this is unlikely if the client
write-opens the FIFO only after receiving ECONNREFUSED, and then does so
immediately, it's still a race condition worth avoiding.)