[DEL] _realpathat.h:  Move constants to unistd.h
Add REALPATH_OUTPUT_* constants to unistd.h (and axe REALPATH_MAC, which
is unused).  _realpathat() is already declared in unistd.h (although its
basis realpath() is in stdlib.h).  This patch removes the constants from
scope for non-relix builds of MacRelix and mac_pathname, but they aren't
used at all in the latter in that case.  For the former, #define them to
-1 so they never match, although system calls are externally unreachable
in a Mach-O MacRelix until we enable loading CFM code, which already has
the prerequisite of redesigning the headers so that relix symbols reside
outside of system header space.