[DEL] uloop:  Axe broken-in-concept unet connector
The problem with uloop is that messages from client to server *and* from
server to client share the same channel.  The only existing peer-to-peer
Freemount program, fping, is as simple as it can be and still can't work
with uloop, because in its counted mode (-c), it shuts down the outgoing
socket for writing after sending the final ping -- and then is unable to
respond with a pong, because that very same socket has already been shut
down.  More complex future programs would have similar issues, and maybe
even worse ones.