macos:  Don't H-rotate patterns based on draw rect
While we sometimes need to offset the pattern vertically (by starting at
a non-zero pattern row index) to correspond to a drawing rectangle whose
top is unaligned (i.e. not a multiple of 8 rows from portRect.top), it's
unnecessary (and wrong) to offset the pattern horizontally, since that's
already accounted for in the left and right edge masks.  Rotation of the
pattern's bits should happen when the port's coordinate system is not in
byte-alignment with the bitmap.