rectangle-math.mac:  Use cc-style test diagnostics
We expect File and Line commands in the diagnostic display of MPW-hosted
tools, and a program like mwcc which knows it interacts with them can be
prepared to deal with such.  Diagnostic-printing tools that run directly
in MacRelix (a Unix-like environment) should use cc-style message lines,
since that's what other Unix tools expect.