rectangle-math.mac:  Write lineref to stderr
Write the file and line information to stderr rather than printing it in
the test case output as the TAP comment.

Under th, it's not displayed anyway, so this is no change.  We also gain
the option of redirecting stderr to a tool for displaying test failures,
in which case it's simpler than parsing the TAP output.  Finally, by not
relying on the TAP comment feature we can eventually remove it entirely.