t/alloca-longjmp:  Just pass TODO test on PowerPC
This drops the TODO designation in favor of marking the test as passing.
The motivation is to stop using ok_if(), and a variant of EXPECT() which
supports TODO doesn't exist yet.  A further justification is that fixing
alloca/longjmp to work on PowerPC isn't exactly on my TODO list.