plus-tests:  Axe capacity tests for 64-bit compat
Tests of string::capacity() are dependent on the structure size, as well
as the details of exactly how much memory to allocate for a given string
length.  Instead of maintaining these tests for both conventional 32-bit
and 64-bit systems (as well as 64-bit Windows, potentially), in addition
to accounting for any implementation changes for performance, axe all of
the capacity tests (and rename the unit appropriately).