v68k:  Allow MOVEP to access odd addresses
MOVEP is special in being either word-sized or long-sized and yet always
accessing memory a byte at a time, so that odd memory addresses aren't a
problem.  Use the high bit in the op params size field to code for this.

This makes the movep.68k tests pass in xv68k.