macos:  Offset portBits bounds in MovePortTo()
Inside Macintosh doesn't explain this, only that "MovePortTo changes the
position of the current grafPort's portRect.  This does *not* affect the
screen; it merely changes the location [of] subsequent drawing".  In any
case, testing reveals that QuickDraw moves *portBits.bounds* so that its
top left is (-top,-left) and doesn't change the portRect whatsoever.