Genie:  Throw EBADF if no read perm for Mac files
Since we're forced to specify read permission when opening a Mac OS file
(because fsWrPerm is only guaranteed to work for drivers), we can't rely
on the Device Manager to check permissions for us for read operations as
it does for writes.  Therefore, explicitly check in the pread method for
read permission in the vfs filehandle flags.