vfs:  Fail open() if truncating but not writing
If the flags include O_TRUNC but the access mode is neither O_WRONLY nor
O_RDWR, throw EINVAL.

We would expect the truncate() call below to fail with EBADF anyway, but
that only happens after the open method returns, and we'd prefer to fail
early, before achieving partial success.  Also, defense in depth demands
that we place a centralized check here rather than relying on the seteof
methods to all do the right thing.

Unix systems tend to perform the truncation (provided write access), but
we're not going to do this.  One, it's awkward to implement (requiring a
separate pathname resolution for the truncation, since the open() result
is read-only and can't be used), and two, there's simply no reason you'd
*ever* want to open a truncated file for reading.  Even if a purpose did
exist, it's explicitly undefined in POSIX -- so it's unportable anyway.