vfs:  Default to node parent's user, if present
If the node's parent is specified upon construction (i.e. as the `owner`
parameter), copy that node's user.  Otherwise, continue to default to 0.

This causes the root node's user to be transitively inherited via almost
all lookups.  The exceptions are the null node and anonymous nodes (i.e.
the result of calling get_file() on a filehandle obtained from pipe() or
socketpair()), where the user doesn't matter because there's no means of
accessing the file system, and HFS nodes that aren't created via lookup,
which like all HFS nodes are currently assumed not to be exported out of
MacRelix, where the user is assumed to be privileged anyway.  (Note that
freemountd uses posix nodes even in MacRelix -- it also links to its own
userspace instance of vfs which lacks HFS nodes.)